% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ao3.scraper-package.R
\docType{package}
\name{ao3.scraper-package}
\alias{ao3.scraper}
\alias{ao3.scraper-package}
\title{ao3.scraper: AO3 Web Scraper in R}
\description{
Scrapes and parses archiveourown.org's webpages to create personal datasets.
}
\author{
\strong{Maintainer}: Pat Thompson \email{patthomp.xyz@gmail.com}

}
\keyword{internal}
