% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collate_pages.R
\name{collate_pages}
\alias{collate_pages}
\title{Collate extracted AO3 pages into a dataset}
\usage{
collate_pages(
  s,
  start_page = NULL,
  end_page = NULL,
  url,
  seconds_to_wait = 5,
  wd
)
}
\arguments{
\item{s}{An rvest session}

\item{start_page, end_page}{Which number page to start/end pulling from}

\item{url}{The general URL to paginate through.}

\item{seconds_to_wait}{How many seconds to wait in between web scrapes.}

\item{wd}{The directory in which to save the result.}
}
\value{
A clean dataset of the URL from \code{start_page} to \code{end_page}
}
\description{
Collate extracted AO3 pages into a dataset
}
