% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_work_ids.R
\name{grab_work_ids}
\alias{grab_work_ids}
\title{Get work ids}
\usage{
grab_work_ids(s, url)
}
\arguments{
\item{s}{Rvest session}

\item{url}{AO3 URL}
}
\value{
A numeric vector of AO3 work ids
}
\description{
Get work ids
}
