% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{login}
\alias{login}
\title{Login to AO3 programmatically}
\usage{
login(user, pw)
}
\arguments{
\item{user}{Your AO3 username.}

\item{pw}{Your AO3 password}
}
\value{
A virtual "logged in" AO3 web session.
}
\description{
Login to AO3 programmatically
}
