% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{population}
\alias{population}
\alias{get_fandoms}
\alias{start_fandom_pull}
\alias{test_for_fandom_size}
\title{Pull population-level data}
\usage{
get_fandoms(s, url, media)

start_fandom_pull(s)

test_for_fandom_size(s, url)
}
\arguments{
\item{s}{An active rvest session.}

\item{url}{URL to webscrape}

\item{media}{AO3 media category for fandoms.}
}
\description{
Pull population-level data
}
\section{Functions}{
\itemize{
\item \code{start_fandom_pull()}: start_fandom_pull

\item \code{test_for_fandom_size()}: Check whether url returns more than 5000 pages (100k works)

}}
