% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_page.R
\name{process_page}
\alias{process_page}
\title{Process raw AO3 link}
\usage{
process_page(s, page_number, url, wd)
}
\arguments{
\item{s}{Rvest session containing the AO3 URL to pull data from, and must
be one of the following:}

\item{page_number}{The numeric page number queried in \code{s}.}

\item{url}{URL of html document to be scraped}

\item{wd}{Directory to which raw JSON should be stored}
}
\description{
Process raw AO3 link
}
