% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_atomics.R
\name{pull_chapters}
\alias{pull_chapters}
\alias{pull_complete_flag}
\alias{pull_summary}
\alias{pull_tags}
\alias{pull_visits}
\alias{pull_warnings}
\title{Extract work-level information from html text document}
\usage{
pull_chapters(text2)

pull_complete_flag(text2)

pull_summary(text2)

pull_tags(text2)

pull_visits(text2)

pull_warnings(text2)
}
\arguments{
\item{text2}{A character string made from converting a webpage into a html document.}
}
\value{
Atomic vectors
}
\description{
Extract work-level information from html text document
}
\section{Functions}{
\itemize{
\item \code{pull_chapters()}: Pull number of chapters

\item \code{pull_complete_flag()}: Pull boolean completeness flag

\item \code{pull_summary()}: Pull author's summary

\item \code{pull_tags()}: Pull all tags e.g.(character, warnings, pairings, etc)

\item \code{pull_visits()}: Pull number of times visited from history page

\item \code{pull_warnings()}: Pull all warning tags

}}
