% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_page.R
\name{pull_page}
\alias{pull_page}
\title{Extract work-level information from html text document}
\usage{
pull_page(tag_ls)
}
\arguments{
\item{tag_ls}{A list containing an html document parsed with html_text() as \code{all_text1},
and the same document parsed with html_text2() as \code{all_text2}.}
}
\value{
Cleaned data.table representing an AO3 page.
}
\description{
Extract work-level information from html text document
}
