% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_api_response.R
\name{save_api_response}
\alias{save_api_response}
\title{Save API response}
\usage{
save_api_response(df, filename, wd)
}
\arguments{
\item{df}{R object to save to a file}

\item{filename}{File name to save the data to}

\item{wd}{Directory to save the .Rda file to}
}
\description{
Save API response
}
