% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_page_raw.R
\name{scrape_page_raw}
\alias{scrape_page_raw}
\title{Scrape a raw AO3 page}
\usage{
scrape_page_raw(s, page_number, wd, url)
}
\arguments{
\item{s}{Rvest session containing the AO3 URL to pull data from, and must
be one of the following:}

\item{page_number}{The numeric page number queried in \code{s}.}

\item{wd}{Directory to which raw JSON should be stored}

\item{url}{URL of html document to be scraped}
}
\value{
List containing the same XML document parsed with html_text and html_text2
}
\description{
Scrape a raw AO3 page
}
